Joining Script + String Splitter
Create script which prevents chatters that haven’t joined by adding those who have joined to a array and checks it every time someone posts a message.
I have also added a bool variable which can be toggled to change whether players need to join to participate or can-do commands without joining.
I have created a basic string splitter that can split a string using a string in put and a char that determines when a word ends.
I have applied the join script to the demo game manager along with the string splitter.