Level related code given own script and level transition completed
All code relating to the control of levels has been moved to the ‘LevelController’ this includes: playerState array
BallHole
ResetBallState sets all values in the playerState to false
BallControl new functions:
FindBallIndex will find the index location of a ball given a game object, it returns the index location if it is part of the balls array or -1 if no ball exists.
HideBall sets the ball to be inactive
MoveBall moves all the balls to the start point of the current level and actives them all
‘End’ prefabs require a link the LevelController in the inspector
When all players have reached the hole they are all moved to the next hole using MoveBall, currently at the last hole the players are put back to the beginning of the last level.