Implementation for MultiTargetCamera
Added interfaces for getting and setting targets.

Added script to be attached to camera: Contains list of GameObjects to target and get transforms from each. Able to add and remove targets during runtime.

Added script to be attached to GameObjects that are to be targetted:
Objects add themselves to the targetting list when Awake and remove themselves OnDisable.

Added enums for tagging different types of target (currently not used)

Created testing scene for Camera Implementation.

Created 2 prefabs for testing purposes: Camera with the camera scripts attached and a Cube with the new targetting request script attached.

Fixed issue where camera movement wasn't working as intended when used at an angle.