Using Magnitude Loss to Stop the Ball
Added a new bool variable m_inMotion which is toggled when it is hit or stopped.
If the ball is in motion and the loss of magnitude is less, then 0.01 then it stops the ball.
Method SetMovement is triggered 3 seconds after the ball is hit and sets m_inMotion to true.