Minor UI Update
Added a player has left the game message on the GUI
Changed the object name in the UI, from "WelcomePlayer" to "PlayerMessage", it can now be used for other messages and can be more universal becuase of the name, making it easier to use it in the future.
Changed the player name display over player ball, currently it shows first 3 letters of the twitch username. Changes to be made to ensure that player name diplay is correct, ensure that if there are 2 or more players with the same name beginning that there is something in place to fix this.