Scene adjustments
Adjusted parts of the existing scene.
Temp: Collision inside hole removed. Ball was occasionally bouncing off of edge of hole at an unexpected angle.
To consider: remodel the hole so this is no longer an issue and to also be more lienent to the player by having bevelled edges (so that the ball can roll in if close enough).