Added Remove Admin
Added method to remove admins and update the XML. You cannot remove  the user set in the playerprefs and you cannot remove non admins.
Also added a method to return the index that the admin is in the admin list. Returns -1 for non admin, -2 for playerpref user and anything greater than or equal to 0 is an admin.