Edited the LevelController script for sounds
Edited the LevelController script so that when a player completes a course it plays the specified sound. Also added a sound which plays when the course is completed by all players. There is also an addition in the script which finds objects with the tag "Player" and adds them to an array; this is so that if there are 2 or more players in-game, and only one player is left to complete the course then it would only play the sound of the course being completed than to have 2 sounds playing at the same time.