Edited Level Controller script for sound
Edited the Level Controller script for the sounds: PlayerCompletedCourse and CourseCompleted.

Removed the unnecessary scripting for playing the sounds based on how many players are in the scene, etc.
Added a function that plays the CourseCompleted sound after 0.5 seconds.
This function is used when all players have completed the current course