Added Max Players Method and Array Shuffle Method
Added method to change the maximum number of players by temporarily storing it and creating a new one. If the transition is possible the method returns true otherwise it returns false.
Added a method to shuffle all users down the array so that all players are next to each other in the array.
Added a method to update minigames with the new list.